Excluding Core Folder and Examples Folder docs
These folders should be excluded because the documents show up blank. Why? Because in the typedocs config internals are excluded. I marked the multiplatform lib as internal. So those pages will be blank. Excluding them will make those pages not be included in the next build

This is the internal tag on C
/** @internal */
export const C =