Blockfrost: Only return empty UTxO list on 404
[Blockfrost documentation](https://docs.blockfrost.io/#tag/Cardano-Addresses/paths/~1addresses~1{address}~1utxos~1{asset}/get)
states that the /utxos/{asset} endpoint can return several different
typs of error codes (e.g., 403, 418, 429) that do not indicate the
address is empty but that the API key is misconfigured or another error
occurred.  This change stops the error from being swallowed in this case
so dApp developers do not get odd behavior when they were expecting an
empty list to indicate truly no UTxOs available.

[ Documentation: None ]
[ Testing: deno task build, Hit blockfrost API using cURL to validate error codes ]