fix(build): Pull from hosted CDN instead of local
Ideally we would want to use the paradigm that the build process set out
to pull from a local lib, but we can't for NextJS.  Since this is a
fork, we don't need to worry about a circular dependency (the fork will
pull from the main).  Without this, NextJS splits the libraries into
chunks and the path resolution fails.

[ Testing: npm i lucid-cardano-nextjs@0.10.10-b4 ]